--- a/rational.js
+++ b/rational.js
@@ -94,6 +94,9 @@ Rational.prototype = {
     isZero: function() {
         return this.p.isZero()
     },
+    isOne: function() {
+        return this.p.equals(1) && this.q.equals(1)
+    },
     isInteger: function() {
         return this.q.equals(bigInt.one)
     },
@@ -139,6 +142,10 @@ Rational.prototype = {
         )
     },
     mul: function(other) {
+        if (this.isZero()) return zero
+        if (other.isZero()) return zero
+        if (this.isOne()) return other
+        if (other.isOne()) return this
         return new Rational(
             this.p.times(other.p),
             this.q.times(other.q)