@terite 2024-10-23
This looks like it's no longer necessary as these changes are included in `master`.